+++ foxy_farmer/farmer/harvester_binary_farmer.py
@@ -28,7 +28,7 @@ class HarvesterBinaryFarmer(Farmer, ABC):
             if run_harvester:
                 await self._binary_environment.start()
             futures: List[Awaitable] = [self._stop_event.wait()]
-            if self._farmer_config.get("monitor_farmer_connections") is True:
+            if self._farmer_config.get("monitor_farmer_connections", True) is True:
                 futures.append(monitor_farmer(root_path=self._root_path, until=self._stop_event))
             await gather(*futures)
         finally: