+++ foxy_farmer/farmer/split_chia_farmer.py
@@ -35,7 +35,7 @@ class SplitChiaFarmer(Farmer, ABC):
             await self._embedded_environment.start_services(self._services_to_run_on_embedded_env)
             await daemon_environment.start_services(self._services_to_run_on_env)
             futures: List[Awaitable] = [self._stop_event.wait()]
-            if self._farmer_config.get("monitor_farmer_connections") is True:
+            if self._farmer_config.get("monitor_farmer_connections", True) is True:
                 futures.append(monitor_farmer(root_path=self._root_path, until=self._stop_event))
             await gather(*futures)
         finally: